refactor: extract reducer initialization to reducer.init.js file
With this change we can use reducer module without initialize the redux store.
